+++ boot/acp-spring-boot-starter/src/main/kotlin/pers/acp/spring/boot/exceptions/SocketException.kt
@@ -1,0 +1,5 @@
+package pers.acp.spring.boot.exceptions
+
+import pers.acp.core.base.BaseException
+
+class SocketException constructor(message: String?) : BaseException(message)